+++ Src/Private/Get-AbrHRZInstantClone.ps1
@@ -5,7 +5,7 @@ function Get-AbrHRZInstantClone {
     .DESCRIPTION
         Documents the configuration of VMware Horizon in Word/HTML/XML/Text formats using PScribo.
     .NOTES
-        Version:        1.1.4
+        Version:        1.1.5
         Author:         Chris Hildebrandt, Karl Newick
         Twitter:        @childebrandt42, @karlnewick
         Editor:         Jonathan Colon, @jcolonfzenpr